requirements.txt
web3==6.10.0
alembic==1.12.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
aiohttp==3.8.4
sqlalchemy==2.0.15